ethtool: add nlchk for redirecting to netlink
To support commands which differ from the ioctl implementation
add a new callback which can check if the arguments on the command
line indicate that the request should be sent over netlink.
The decision should be inferred from the arguments, rather
than an explicit --netlink argument.

v3: fix --disable-netlink build

Signed-off-by: Jakub Kicinski <kuba@kernel.org>