netlink: eeprom: Fallback to IOCTL when a complete hex/raw dump is requested
The IOCTL backend provides a complete hex/raw dump of the module EEPROM
contents:

 # ethtool -m swp11 hex on | wc -l
 34

 # ethtool -m swp11 raw on | wc -c
 512

With the netlink backend, only the first 128 bytes from I2C address 0x50
are dumped:

 # ethtool -m swp11 hex on | wc -l
 10

 # ethtool -m swp11 raw on | wc -c
 128

The presence of optional / banked pages is unknown without parsing the
EEPROM contents which is unavailable when pretty printing is disabled
(i.e., configure --disable-pretty-dump). With the IOCTL backend, this
parsing happens inside the kernel.

Therefore, when a complete hex/raw dump is requested, fallback to the
IOCTL backend.

After the patch:

 # ethtool -m swp11 hex on | wc -l
 34

 # ethtool -m swp11 raw on | wc -c
 512

This avoids breaking users that are relying on current behavior.

If users want a hex/raw dump of optional/banked pages that are not
returned with the IOCTL backend, they will be required to request these
explicitly via the netlink backend. For example:

 # ethtool -m swp11 hex on page 0x2

This is desirable as that way there is no ambiguity regarding the
location of optional/banked pages in the dump.

Another way to implement the above would be to use the 'nlchk' callback
added in commit 67a9ef551661 ("ethtool: add nlchk for redirecting to
netlink"). However, it is called before the netlink instance is
initialized and before the command line parameters are parsed via
nl_parser().

Fixes: 25b64c66f58d ("ethtool: Add netlink handler for getmodule (-m)")
Signed-off-by: Ido Schimmel <idosch@nvidia.com>