sff-8636: Consolidate code between IOCTL and netlink paths
Now that both the netlink and IOCTL paths use the same memory map
structure for parsing, the code can be easily consolidated.

Note that the switch-case statement is not necessary for the netlink
path, as the netlink code (i.e., netlink/module-eeprom.c) already
performed the check, but it is required for the IOCTL path.

Signed-off-by: Ido Schimmel <idosch@nvidia.com>