ethtool: Set mask correctly for dumping advertised FEC modes
Set the "mask" value to false when dumping the advertised FEC modes.
The advertised values are stored in the value field, while the supported
values are what is stored in the mask.

Without this change the supported value is displayed for both the supported
and advertised modes resulting in the advertised value being ignored.

Signed-off-by: Alexander Duyck <alexanderduyck@fb.com>